feat(maps): Add Countries Choropleth
