+++ packages/map/index.html
@@ -1,0 +1,59 @@
+<!doctype html>
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <title>@hpcc-js/chart</title>
+
+    <script>
+        if (window.location.protocol === "file:" || window.location.origin.indexOf("http://localhost") === 0) {
+            document.write('<link rel="stylesheet" href="../../packages/common/font-awesome/css/font-awesome.min.css">');
+            document.write('<script type="text/javascript" src="../../packages/loader/dist/index.js"><' + '/script>');
+        } else {
+            document.write('<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@hpcc-js/common/font-awesome/css/font-awesome.min.css">');
+            document.write('<script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@hpcc-js/loader"><' + '/script>');
+        }
+    </script>
+
+    <script>
+        var hpccLoader = window["@hpcc-js/loader"];
+    </script>
+
+    <style>
+        #placeholder,
+        .placeholder {
+            position: absolute;
+            left: 8px;
+            top: 8px;
+            right: 8px;
+            bottom: 8px;
+        }
+    </style>
+</head>
+
+<body onresize="doResize()">
+    <div id="placeholder">
+    </div>
+    <script>
+        var app;
+        hpccLoader.amd().then(function (require) {
+            require(["lib-umd/test"], function (testMod) {
+
+                app = new testMod.Test()
+                    .target("placeholder")
+                    .render();
+
+            });
+        });
+
+        function doResize() {
+            if (app) {
+                app
+                    .resize()
+                    .lazyRender();
+            }
+        }
+    </script>
+</body>
+
+</html>
